requirements.txt
uvicorn~=0.40.0
python-dotenv~=1.2.1
fastapi~=0.128.0
starlette~=0.50.0
requests~=2.32.5
librouteros~=3.4.1
cryptography~=46.0.3
bcrypt~=5.0.0
aiohttp~=3.13.3
pydantic~=2.12.5
paramiko~=4.0.0